[fix] tooltip.js에서 placementMap, placements 공통 상수 export
