+++ src/utils/constants/tooltip.js
@@ -1,0 +1,22 @@
+export const placements = [
+	'top',
+	'top-right',
+	'top-left',
+	'bottom',
+	'bottom-right',
+	'bottom-left',
+	'right-top',
+	'right',
+	'right-bottom',
+];
+export const placementMap = {
+	top: 'top',
+	'top-right': 'top-end',
+	'top-left': 'top-start',
+	bottom: 'bottom',
+	'bottom-right': 'bottom-end',
+	'bottom-left': 'bottom-start',
+	right: 'right',
+	'right-top': 'right-start',
+	'right-bottom': 'right-end',
+};
